- Upgraded premake5 from 5.0.0-alpha15 to 5.0.0-alpha16 - Changed Logo - Removed compress option, Now all encryptions being compressed by default - Added logo image in mainmenu scene

+++ Source/Utility/Random.hpp
@@ -60,17 +60,17 @@ public:
 		String str(length, '\000');
 		for (char& c : str)
 		{
-			const i16 i = Int(static_cast<i16>(1), static_cast<i16>(4));
+			const i32 i = Int(static_cast<i32>(1), static_cast<i32>(4));
 			switch (i)
 			{
 			case 1: // a-z
-				c = Int(static_cast<i16>('a'), static_cast<i16>('z'));
+				c = Int(static_cast<i32>('a'), static_cast<i32>('z'));
 				break;
 			case 2: // A-Z
-				c = Int(static_cast<i16>('A'), static_cast<i16>('Z'));
+				c = Int(static_cast<i32>('A'), static_cast<i32>('Z'));
 				break;
 			case 3: // 0-9
-				c = Int(static_cast<i16>('0'), static_cast<i16>('9')); // ascii 0 = 48 | 9 = 57
+				c = Int(static_cast<i32>('0'), static_cast<i32>('9')); // ascii 0 = 48 | 9 = 57
 				break;
 			case 4: // special characters
 				using Constants::Algorithm::SPECIAL_CHARACTERS;
